+++ src/main/java/es/javier/timerPlugin/configs/TimerConfig.java
@@ -1,0 +1,13 @@
+package es.javier.timerPlugin.configs;
+
+public  class TimerConfig {
+    private final int totalSeconds;
+
+    public TimerConfig(int totalSeconds) {
+        this.totalSeconds = totalSeconds;
+    }
+
+    public int getTotalSeconds() {
+        return totalSeconds;
+    }
+}
